+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated `Strings::Generate` to function without length parameter.
 - Implemented `Primitives::GenerateSeq` overload for int, double and string.
 - Fixed `EnumerableExtensions::Choice` extension method.
+- Refactored `ArbitraryValue` to generate default values.
 
 ## [6.2.1] - 2023-05-26